requirements.txt
aiohttp==3.8.3
aiohttp-jinja2==1.5.1
beautifulsoup4==4.11.2
Flask==2.0.3
pillow
segmentation-models-pytorch
numpy
opencv-python==4.6.0.66
